Renaming license-checker to license
