+++ server/license/checker.go
