+++ server/license/checker_test.go
